+++ Dionysos/Services/InventoryItemServices/InventoryItemDeletionService.cs
@@ -13,8 +13,11 @@ public class InventoryItemDeletingService
 
     public void DeleteInventoryItem(int id)
     {
-        var item = _dbContext.InventoryItems.Single(x => x.Id == id);
-        _dbContext.InventoryItems.Remove(item);
+        var item = _dbContext.InventoryItems.SingleOrDefault(x => x.Id == id);
+        if (item is not null)
+        {
+            _dbContext.InventoryItems.Remove(item);
+        }
         _dbContext.SaveChanges();
     }
 }
