+++ Dionysos/Services/VendorServices/VendorDeletionService.cs
@@ -13,8 +13,11 @@ public class VendorDeletionService
 
     public void DeleteVendor(int id)
     {
-        var vendor = _dbContext.Vendors.Single(x => x.Id == id);
-        _dbContext.Vendors.Remove(vendor);
+        var vendor = _dbContext.Vendors.SingleOrDefault(x => x.Id == id);
+        if (vendor is not null)
+        {
+            _dbContext.Vendors.Remove(vendor);
+        }
         _dbContext.SaveChanges();
     }
 }
